[ protect Video Edit get and post, Delete ] owner가 아닌 경우 비디오 파일을 수정 또는 삭제 할 수 없도록 백앤드에서의 보호도 꼭 필요하다! 해당 작업의 경우처럼 id값만 가져와 작업해도 충분할 경우 populate를 이용해 작업 할 필요가 없다. populate를 이용할 때와 이용하지 않을 때를 잘 구분하여 필요없는 작업을 하지 않도록 주의하여 사용하자.

+++ src/controllers/videoController.js
@@ -12,6 +12,7 @@ export const home = async (req, res) => {
 };
 export const watch = async(req, res) => {
     const {id} = req.params;
+    
     const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found"});    
@@ -20,18 +21,30 @@ export const watch = async(req, res) => {
 }
 export const getEdit = async(req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
+
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found"})
     }
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
+
     return res.render("edit", {pageTitle: `[Edit] ${video.title}`, video });
 };
 export const postEdit = async(req, res) => {
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id:id});
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
     if(!video){
         return res.status(404).redirect("404", {pageTitle: `Video not found`});
+    }
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -78,7 +91,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found"})
+    };
+
+    if(String(_id) !== String(video.owner)){
+        return res.status(403).redirect("/");
+    };
+
     await Video.findByIdAndRemove(id);
+
     return res.redirect("/");
 }
